Huckleberry Finn

+++ ENUM_COM.h
@@ -1,4 +1,4 @@
-#define CMD_CMD(NAME, COD, ARG,...) CMD_ ## NAME = (int)COD, 
+#define CMD_CMD(NAME, COD, ARG,...) CMD_ ## NAME = COD, 
 enum COMAND 
 {
     #include "commands.h"
